Fixed Middleware
re structure the endpoints
