+++ src/router/chat.ts
@@ -1,0 +1,9 @@
+import express from 'express';
+
+import { getChatMessages } from '../controllers';
+import { isAuthenticated  } from '../middlewares';
+
+export default (router: express.Router) => {
+    router.get('/getMessages',isAuthenticated, getChatMessages);
+
+}
