+++ src/router/ticketing.ts
@@ -1,0 +1,9 @@
+import express from 'express';
+
+import { submitDocument } from '../controllers';
+import { isAuthenticated  } from '../middlewares';
+
+export default (router: express.Router) => {
+    router.post('/submitForm',isAuthenticated, submitDocument);
+
+}
