+++ src/types/express.d.ts
@@ -1,0 +1,8 @@
+// express.d.ts
+import express from 'express';
+
+declare module 'express-serve-static-core' {
+    interface Request {
+        sessionToken?: string;
+    }
+}
